refactor(pages): share SectionHeading and CSV helpers across directory pages
