+++ src/pages/links.tsx
@@ -1,7 +1,9 @@
-import React, { JSX, ReactNode } from 'react';
+import React, { JSX } from 'react';
 import Layout from '@theme/Layout';
 import Head from '@docusaurus/Head';
 import LinksCard from '@site/src/components/LinksCard'; // --- news page と共通化 ---
+import SectionHeading from '@site/src/components/SectionHeading';
+import { escapeCsvField, downloadCsv } from '@site/src/lib/csv';
 import styles from './links.module.css';
 import Translate, { translate } from '@docusaurus/Translate';
 import { Download } from 'lucide-react';
@@ -166,37 +168,7 @@ const linksData: LinkCategoryData[] = [
   },
 ];
 
-/**
- * ホバー時に右側にアンカーリンク('#')を表示する見出しコンポーネント
- */
-const SectionHeading = ({ as: Component, id, className, children }: { as: 'h2' | 'h3'; id: string; className: string; children: ReactNode }) => (
-  // h2(Component)に行全体をホバー領域とするためのクラスを適用
-  <Component id={id} className={className}>
-    {/* テキストとアンカーリンクをspanで囲み、アンカーの位置決めの基準とする */}
-    <span className={styles.anchorContainer}>
-      {children}
-      <a
-        className={styles.anchorLink}
-        href={`#${id}`}
-        aria-label={translate({
-          id: 'theme.common.headingLinkTitle',
-          message: 'この見出しへの固定リンク',
-        })}>
-        #
-      </a>
-    </span>
-  </Component>
-);
-
-
 export default function LinksPage(): JSX.Element {
-  const escapeCsvField = (field: string): string => {
-    if (/[",\n]/.test(field)) {
-      return `"${field.replace(/"/g, '""')}"`;
-    }
-    return field;
-  };
-
   const handleDownload = () => {
     const header = [
       translate({ id: 'links.download.header.siteName', message: 'SiteName' }),
@@ -210,14 +182,7 @@ export default function LinksPage(): JSX.Element {
       return [escapeCsvField(siteTitle), site.href].join(',');
     });
 
-    const csvContent = [header, ...rows].join('\n');
-    const blob = new Blob([`\uFEFF${csvContent}`], { type: 'text/csv;charset=utf-8;' });
-    const link = document.createElement('a');
-    link.href = URL.createObjectURL(blob);
-    link.setAttribute('download', 'hkdocs-links-sites.csv');
-    document.body.appendChild(link);
-    link.click();
-    document.body.removeChild(link);
+    downloadCsv('hkdocs-links-sites.csv', header, rows);
   };
 
   return (
